+++ resources/views/layouts/admin/aside.blade.php
@@ -56,6 +56,7 @@
                             <x-aside.dropdownlink :label="__('Allgemein')" href="{{ route('admin.general.index') }}" />
                             <x-aside.dropdownlink :label="__('Fernwartung')" href="{{ route('admin.setting.index') }}" />
                             <x-aside.dropdownlink :label="__('Mail')" href="{{ route('admin.mail.index') }}" />
+                            <x-aside.dropdownlink :label="__('Sicherheit')" href="{{ route('admin.security.index') }}" />
                         @endcan
                         {{-- Die Aufbewahrungsfrist gehoert zum Protokoll, nicht zu
                              den uebrigen Einstellungen. --}}